1. move W_Dynamic from object.py into primitive.py
2. we finally can remove the special case about W_Dynamic in
get_spy_type: we can now use the "official" Annotated capability of
@builtin_func
